refactor: split trade market row header and controls parts (#3878)
Extract _MarketCommodityRowHeader and _MarketCommodityRowControls into
dedicated library parts so the trade-screen market row host stays under
the repo file-size target; document the new parts in repo-and-packages.md.

Refs #3878
