+++ SPEC/program/repo-and-packages.md
@@ -60,7 +60,7 @@ Monorepo layout:
 - **`features/game/screens/technology/`** — Technology screen library ([TechnologyScreen](../../app/lib/features/game/screens/technology/technology_screen.dart)). Public imports use full package paths; no barrel re-export shim.
 - **`features/game/widgets/chrome/`** — Ct-* catalog widget implementations shared by in-game panels and dialogs: hover button mixin ([ct_hover_button.dart](../../app/lib/features/game/widgets/chrome/ct_hover_button.dart)), nine-patch button re-export shim ([ct_nine_patch_button.dart](../../app/lib/features/game/widgets/chrome/ct_nine_patch_button.dart) → source of truth [app/lib/widgets/ct_nine_patch_button.dart](../../app/lib/widgets/ct_nine_patch_button.dart) with state/brackets `part`s), action/danger text buttons, dialog shell, panel, and circular locate button. Flame `GameWidget` embedding lives in `widgets/shell/region_map_game_viewport.dart` so this folder stays free of direct Flame imports in catalog widgets.
 - **`features/game/widgets/dialogue/`** — narrative dialogue overlays and shared view chrome: [CtDialogueView](../../app/lib/features/game/widgets/dialogue/ct_dialogue_view.dart) + line/choice body parts, game-start intro ([game_start_intro_overlay.dart](../../app/lib/features/game/widgets/dialogue/game_start_intro_overlay.dart)), overture overlay host + yarn-flow and offer-row `part`s ([overture_dialogue_overlay.dart](../../app/lib/features/game/widgets/dialogue/overture_dialogue_overlay.dart), [overture_dialogue_overlay_flow.dart](../../app/lib/features/game/widgets/dialogue/overture_dialogue_overlay_flow.dart), [overture_dialogue_overlay_offer_row.dart](../../app/lib/features/game/widgets/dialogue/overture_dialogue_overlay_offer_row.dart)), intervention overlay + yarn-flow `part` ([intervention_dialogue_overlay.dart](../../app/lib/features/game/widgets/dialogue/intervention_dialogue_overlay.dart), [intervention_dialogue_overlay_flow.dart](../../app/lib/features/game/widgets/dialogue/intervention_dialogue_overlay_flow.dart)) and extracted choice buttons ([intervention_choice_buttons.dart](../../app/lib/features/game/widgets/dialogue/intervention_choice_buttons.dart)), call-to-arms overlay host + call-row `part` ([call_to_arms_dialogue_overlay.dart](../../app/lib/features/game/widgets/dialogue/call_to_arms_dialogue_overlay.dart), [call_to_arms_call_row.dart](../../app/lib/features/game/widgets/dialogue/call_to_arms_call_row.dart)), tribe first-contact overlay + sync helper. Public imports use full package paths; no barrel re-export shim.
-- **`features/game/screens/trade/`** — World Market trade screen library ([TradeScreen](../../app/lib/features/game/screens/trade/trade_screen.dart) with `part` files for widget-test contract aggregator (`trade_screen_contract.dart`), Market contract keys (`trade_screen_contract_market.dart`), Deal Book contract keys (`trade_screen_contract_deal_book.dart`), two-tab body chrome (`trade_screen_tabs_body.dart`), Market tab host (`trade_screen_market_tab.dart`), Market tab `build` implementation (`trade_screen_market_tab_build.dart`), cargo header (`trade_screen_market_tab_cargo_header.dart`), order handlers (`trade_screen_market_tab_order_handlers.dart`), catalog/section builders (`trade_screen_market_tab_catalog.dart`), Market row chrome (`trade_screen_market_row.dart`, `trade_screen_market_row_stepper.dart`), and Deal Book ledger (`trade_screen_deal_book.dart`, `trade_screen_deal_book_panel.dart`, `trade_screen_deal_book_rows.dart`)), plus shared Market-tab section callback factory ([trade_section_handlers.dart](../../app/lib/features/game/screens/trade/trade_section_handlers.dart)). Public imports use full package paths; no barrel re-export shim.
+- **`features/game/screens/trade/`** — World Market trade screen library ([TradeScreen](../../app/lib/features/game/screens/trade/trade_screen.dart) with `part` files for widget-test contract aggregator (`trade_screen_contract.dart`), Market contract keys (`trade_screen_contract_market.dart`), Deal Book contract keys (`trade_screen_contract_deal_book.dart`), two-tab body chrome (`trade_screen_tabs_body.dart`), Market tab host (`trade_screen_market_tab.dart`), Market tab `build` implementation (`trade_screen_market_tab_build.dart`), cargo header (`trade_screen_market_tab_cargo_header.dart`), order handlers (`trade_screen_market_tab_order_handlers.dart`), catalog/section builders (`trade_screen_market_tab_catalog.dart`), Market row chrome (`trade_screen_market_row.dart`, `trade_screen_market_row_header.dart`, `trade_screen_market_row_controls.dart`, `trade_screen_market_row_stepper.dart`), and Deal Book ledger (`trade_screen_deal_book.dart`, `trade_screen_deal_book_panel.dart`, `trade_screen_deal_book_rows.dart`)), plus shared Market-tab section callback factory ([trade_section_handlers.dart](../../app/lib/features/game/screens/trade/trade_section_handlers.dart)). Public imports use full package paths; no barrel re-export shim.
 - **`features/game/screens/combat/`** — Quick Battle full-screen host ([QuickBattleScreen](../../app/lib/features/game/screens/combat/quick_battle_screen.dart)). Public imports use full package paths; no barrel re-export shim.
 - **`features/game/widgets/combat/`** — Quick Battle modal and presentational widgets: combat mode choice ([combat_mode_choice_dialog.dart](../../app/lib/features/game/widgets/combat/combat_mode_choice_dialog.dart)), result summary ([quick_battle_result_dialog.dart](../../app/lib/features/game/widgets/combat/quick_battle_result_dialog.dart)), deployment grid ([quick_battle_deployment_view.dart](../../app/lib/features/game/widgets/combat/quick_battle_deployment_view.dart)), and CP action selector ([quick_battle_action_selector.dart](../../app/lib/features/game/widgets/combat/quick_battle_action_selector.dart)). Opened via `OpenDialogEvent` registrations in `AppEventHandlerScope`. Public imports use full package paths; no barrel re-export shim.
 - **Colocated UI helpers (no `features/game/utils/` barrel):** technology helpers ([research_slot_preview.dart](../../app/lib/features/game/widgets/technology/research_slot_preview.dart), [tech_gp_researchers.dart](../../app/lib/features/game/widgets/technology/tech_gp_researchers.dart), [tech_ui_helpers.dart](../../app/lib/features/game/widgets/technology/tech_ui_helpers.dart)); production commodity labels ([commodity_ui_helpers.dart](../../app/lib/features/game/widgets/production/commodity_ui_helpers.dart)); unit-panel region labels ([region_labels.dart](../../app/lib/features/game/widgets/units/shared/region_labels.dart)); map locate tile keys ([map_location_resolver.dart](../../app/lib/features/game/flame/map_state/map_location_resolver.dart)); sea-zone display names ([sea_zone_name_resolver.dart](../../app/lib/features/game/widgets/province_overlay/sea_zone_name_resolver.dart)).
