+++ app/lib/features/game/screens/trade/trade_screen_market_row.dart
@@ -1,4 +1,4 @@
-// Market tab commodity row widgets for the World Market Trade screen
+// Market tab commodity row host for the World Market Trade screen
 // (Refs #2993, #3093, #3487, split out from `trade_screen.dart` to keep
 // the host file under the repo-lint non-comment line limit per
 // `SPEC/program/dart-file-non-comment-line-size.md`).
@@ -117,178 +117,6 @@ class _MarketCommodityRow extends StatelessWidget {
           onDirectionChanged: onDirectionChanged,
           onIncrement: onIncrement,
           onDecrement: onDecrement,
-        ),
-      ],
-    );
-  }
-}
-
-/// Top read-only line of a Market tab commodity row (Refs #3093, #3487):
-/// resource icon, commodity name, sellable `(N)` headroom readout, then a
-/// fixed-width trailing column that right-aligns the coin icon and price so
-/// every row shares the same price-digit column edge.
-/// Extracted from `_MarketCommodityRow.build` to keep the parent
-/// `build` body within the `widget_build_method_too_long` AST cap.
-class _MarketCommodityRowHeader extends StatelessWidget {
-  const _MarketCommodityRowHeader({
-    required this.commodityId,
-    required this.commodityDisplayName,
-    required this.priceText,
-    required this.sellableHeadroom,
-    required this.nameStyle,
-    required this.priceStyle,
-  });
-
-  final CommodityId commodityId;
-  final String commodityDisplayName;
-  final String priceText;
-  final int sellableHeadroom;
-  final TextStyle nameStyle;
-  final TextStyle priceStyle;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextStyle sellableStyle = nameStyle.copyWith(
-      color: EditorialMonoclePalette.muted,
-    );
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: <Widget>[
-        ResourceIcon(
-          key: TradeScreen.marketRowResourceIconKey(commodityId),
-          commodityId: commodityId,
-          size: TradeScreen.marketRowResourceIconSize,
-        ),
-        const SizedBox(width: 4),
-        Expanded(
-          child: Row(
-            children: <Widget>[
-              Flexible(
-                child: Text(
-                  commodityDisplayName,
-                  style: nameStyle,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-              const SizedBox(width: 4),
-              Text(
-                // ignore: avoid_hardcoded_strings_in_widgets
-                '($sellableHeadroom)',
-                key: TradeScreen.marketRowSellableReadoutKey(commodityId),
-                style: sellableStyle,
-              ),
-            ],
-          ),
-        ),
-        SizedBox(
-          width: TradeScreen.marketRowPriceColumnWidth,
-          child: Align(
-            alignment: Alignment.centerRight,
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: <Widget>[
-                StrictAssetIcon(
-                  key: TradeScreen.marketRowPriceCoinIconKey(commodityId),
-                  assetPath: TradeScreen.marketRowPriceCoinAssetPath,
-                  width: TradeScreen.marketRowPriceCoinIconSize,
-                  height: TradeScreen.marketRowPriceCoinIconSize,
-                ),
-                const SizedBox(
-                  width: TradeScreen.marketRowPriceColumnInnerGap,
-                ),
-                Text(priceText, style: priceStyle),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-/// Interactive controls strip below the static read-only data on a
-/// Market tab commodity row (Refs #2993 E5b).
-class _MarketCommodityRowControls extends StatelessWidget {
-  const _MarketCommodityRowControls({
-    required this.commodityId,
-    required this.stagedType,
-    required this.quantityText,
-    required this.quantityStyle,
-    required this.canDecrement,
-    required this.canIncrement,
-    required this.canSelectOffer,
-    required this.onDirectionChanged,
-    required this.onIncrement,
-    required this.onDecrement,
-  });
-
-  final CommodityId commodityId;
-  final TradeOrderType? stagedType;
-  final String quantityText;
-  final TextStyle quantityStyle;
-  final bool canDecrement;
-  final bool canIncrement;
-
-  /// Refs #3093 — sellable clamp slice. When `false` the Offer chip
-  /// reads as disabled (no `onSelected` handler) so the player cannot
-  /// stage an offer for a commodity whose per-commodity offer cap is
-  /// `0`. Offers already staged on the row remain interactive (the
-  /// player can decrement / drop them) regardless of this gate.
-  final bool canSelectOffer;
-
-  final ValueChanged<TradeOrderType?> onDirectionChanged;
-  final VoidCallback onIncrement;
-  final VoidCallback onDecrement;
-
-  @override
-  Widget build(BuildContext context) {
-    return Wrap(
-      spacing: 6,
-      runSpacing: 6,
-      crossAxisAlignment: WrapCrossAlignment.center,
-      children: <Widget>[
-        CtChoiceChip(
-          key: TradeScreen.marketRowNoneChipKey(commodityId),
-          selected: stagedType == null,
-          onSelected: (_) => onDirectionChanged(null),
-          label: const Text(_MarketTabContent.noneChipLabel),
-        ),
-        CtChoiceChip(
-          key: TradeScreen.marketRowBidChipKey(commodityId),
-          selected: stagedType == TradeOrderType.bid,
-          onSelected: (_) => onDirectionChanged(TradeOrderType.bid),
-          label: const Text(_MarketTabContent.bidChipLabel),
-        ),
-        CtChoiceChip(
-          key: TradeScreen.marketRowOfferChipKey(commodityId),
-          selected: stagedType == TradeOrderType.offer,
-          onSelected: canSelectOffer
-              ? (_) => onDirectionChanged(TradeOrderType.offer)
-              : null,
-          label: const Text(_MarketTabContent.offerChipLabel),
-        ),
-        _StepperButton(
-          buttonKey: TradeScreen.marketRowDecrementKey(commodityId),
-          // ignore: avoid_hardcoded_strings_in_widgets
-          glyph: '−',
-          semanticLabel: _MarketTabContent.decrementSemanticLabel,
-          onPressed: canDecrement ? onDecrement : null,
-        ),
-        SizedBox(
-          width: 28,
-          child: Text(
-            quantityText,
-            key: TradeScreen.marketRowQuantityTextKey(commodityId),
-            style: quantityStyle,
-            textAlign: TextAlign.center,
-          ),
-        ),
-        _StepperButton(
-          buttonKey: TradeScreen.marketRowIncrementKey(commodityId),
-          // ignore: avoid_hardcoded_strings_in_widgets
-          glyph: '+',
-          semanticLabel: _MarketTabContent.incrementSemanticLabel,
-          onPressed: canIncrement ? onIncrement : null,
         ),
       ],
     );
